Extract tag page logic into new file, move atlas constants into new file
