+++ src/build-constants.js
@@ -1,0 +1,5 @@
+// Atlas DB config
+module.exports.DOCUMENTS_COLLECTION = 'documents';
+module.exports.ASSETS_COLLECTION = 'assets';
+module.exports.METADATA_COLLECTION = 'metadata';
+module.exports.SNOOTY_STITCH_ID = 'snooty-koueq';
